refactor validators
